Factorize Chapter info view
